requirements.txt
python-dotenv==0.11.0
notebook==6.0.3
ipykernel==5.1.4
numpy==1.18.1
pandas==1.0.1
scipy==1.4.1
matplotlib==3.1.3
seaborn==0.10.0
scikit-learn==0.22.1
nltk==3.4.5
spacy==2.2.3
gensim==3.8.1
ipywidgets==7.5.1
